% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merged_dataframes.R
\name{add_in.x}
\alias{add_in.x}
\title{Add a constant (TRUE) variable 'in.x' to a data frame.}
\usage{
add_in.x(df)
}
\arguments{
\item{df}{The input data frame.}
}
\value{
The input data frame with the new variable in.x added.
}
\description{
Returns the input data frame after adding the TRUE variable in.x.
}
\details{
Simple wrapper for add_named_constant.
}
\examples{
df.test <- data.frame(v1 = 1:3)
df.test
add_in.x(df.test)
}
\author{
Olivier Humblet
}

