% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merged_dataframes.R
\name{add_named_constant}
\alias{add_named_constant}
\title{Add a constant variable to a data frame.}
\usage{
add_named_constant(df, new_varname, value)
}
\arguments{
\item{df}{The input data frame.}

\item{new_varname}{The name(s) of the new variable(s): a character vector of names, each of which will be assigned to a new variable, all with the same value.}

\item{value}{The single value that will be taken by all the new variables. Can be char, logical, or numeric.}
}
\value{
The input data frame with the new variable(s) added.
}
\description{
Returns the input data frame after adding one (or several) new variable(s) with specified name(s) and a single specified value.
}
\details{
Does not check the type of the value variable.
}
\examples{
df.test <- data.frame(v1 = 1:3)
df.test
add_named_constant(df.test, "in.x", TRUE)
}
\author{
Olivier Humblet
}

