% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_obs_and_ids.R
\name{count_unique_id_combos}
\alias{count_unique_id_combos}
\title{Count data frame rows with unique combinations of the specified variables.}
\usage{
count_unique_id_combos(df, by.i = NULL)
}
\arguments{
\item{df}{The input data frame.}

\item{by.i}{A character vector of the names of by variables for that data frame. For example, by.x if the specified data frame is 'x' and by.x  is specified.}
}
\value{
A single number representing the count of unique rows.
}
\description{
Count data frame rows with unique combinations of the specified variables.
}
\details{
None.
}
\examples{
df.test <- data.frame(v1 = rep(c(1,2,3), c(3,3,3)),
v2 = rep(c(1,2,3), 3),
v3 = runif(9))
df.test

count_unique_id_combos(df.test, "v1")
}
\author{
Olivier Humblet
}

