% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_functions.R
\name{generate_errors}
\alias{generate_errors}
\title{Generate errors.}
\usage{
generate_errors(by, by.x, by.y, list_in)
}
\arguments{
\item{by}{The 'by' input to the merge function.}

\item{by.x}{The 'by.x' input to the merge function.}

\item{by.y}{The 'by.y' input to the merge function.}

\item{list_in}{A list containing 3 named data frames:  merged is the data frame that resulted from merging x and y, merged_from_x is the rows of merged that came from x, and merged_from_y is the rows of merged that came from y.}
}
\value{
None, generates errors.
}
\description{
Generate errors.
}
\details{
Generates errors for the following conditions: no by variables are provided.
}
\examples{
generate_errors(NULL, NULL, NULL, list())
}
\author{
Olivier Humblet
}

