% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/return_join_type.R
\name{return_join_type}
\alias{return_join_type}
\title{Return the join type.}
\usage{
return_join_type(all, all.x, all.y)
}
\arguments{
\item{all}{Identical to base merge function.}

\item{all.x}{Identical to base merge function.}

\item{all.y}{specifications of the columns used for merging (See Details), but differs from base function in that default is NULL, to prevent inadvertent merging without a by variable.}
}
\value{
Returns a character value equal to one of "INNER", "FULL OUTER", "LEFT OUTER", "RIGHT OUTER". Or an error if an unexpected combination of values is specified.
}
\description{
Take as inputs the three 'all' variables, and determines what merge type was intended, i.e. one of "INNER", "FULL OUTER", "LEFT OUTER", "RIGHT OUTER".
}
\details{
Introduces a behavior that differs from that of base merge, i.e. that ir returns an error for combinations of 'all' values that are odd but legal. For example return_merge_type(all = "FALSE", all.x = "TRUE", all.y = FALSE) returns an error because it was probably unintentional, even though this seems to produce a functional merge (presumably a FULL OUTER merge)
}
\examples{
return_merge_type(all = "FALSE", all.x = "FALSE", all.y = FALSE)
}
\author{
Olivier Humblet
}
\keyword{merge}

