% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_obs_and_ids.R
\name{return_list_ratios_xy}
\alias{return_list_ratios_xy}
\title{Returns a list of the ratios of observations to rows in x and y.}
\usage{
return_list_ratios_xy(x, y, by.x, by.y)
}
\arguments{
\item{x}{Typically the merged portion of the final merged dataset that came from x. (We only care about the merged portion.)}

\item{y}{Typically the merged portion of the final merged dataset that came from y. (We only care about the merged portion.)}

\item{by.x}{The 'by.x' input to the merge function.}

\item{by.y}{The 'by.y' input to the merge function.}
}
\value{
Returns a list whose two entries are the ratios of observations to unique by-variable combinations, for the two datasets (i.e. ratio_in_x and ratio_in_y, respectively).
}
\description{
Returns a list of the ratios of observations to rows in both x and y..
}
\details{
Additional details, including preconditions and warnings (which are not supported roxygen sections).
}
\examples{
Examples of using the function.
}
\author{
Olivier Humblet
}

