% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merged_dataframes.R
\name{return_merged_subset}
\alias{return_merged_subset}
\title{Return the subset of a merged data frame from the specified input data frame.}
\usage{
return_merged_subset(dIn, varname)
}
\arguments{
\item{dIn}{The input data frame. It is assumed that dIn was created by merging together two data drames. dIn must contain the boolean constant variable whose name is specified by varname. It is assumed that the varname variable was added to oe of the input data frames, with the value TRUE. Therefore all occurrences of a value TRUE in dIn indicate a row that came from the desired input data frame.}

\item{varname}{The name of the variable in dIn that is descibed above.}
}
\value{
Returns the subset of dIn for which the value of varname is TRUE. This should indicate the rows of dIn that were merged-in the desired input data frame.
}
\description{
Return the subset of a merged data frame from the specified input data frame, based on the presence of a boolean variable.
}
\details{
None.
}
\examples{
d <- data.frame(id = 1:2, from.x = c(TRUE, NA), from.y = NA)
return_merged_subset(d, "from.x")
}
\author{
Olivier Humblet
}

