% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SafeMerge functions.R
\name{SafeMerge}
\alias{SafeMerge}
\title{Safer  merging of two data frames.}
\usage{
SafeMerge(x, y, by = NULL, by.x = by, by.y = by, all = FALSE,
  all.x = all, all.y = all, sort = TRUE, suffixes = c(".x", ".y"),
  incomparables = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Identical to base merge function.}

\item{y}{Identical to base merge function.}

\item{by}{specifications of the columns used for merging (See ‘Details’), but differs from base function in that default is NULL, to prevent inadvertent merging without a by variable.}

\item{by.x}{Identical to base merge function.}

\item{by.y}{Identical to base merge function.}

\item{all}{Identical to base merge function.}

\item{all.x}{Identical to base merge function.}

\item{all.y}{Identical to base merge function.}

\item{sort}{Identical to base merge function.}

\item{suffixes}{Identical to base merge function.}

\item{incomparables}{Identical to base merge function.}

\item{verbose}{Logical; if FALSE then less information is provided.}
}
\value{
Returns the data frame resulting from the merge of the two input data frames.  The functionality is mostly identical to 'merge', except that the by variable is set to NULL by default, no dots parameters can be specified, and there is an additional parameter verbose). See 'Details'.
}
\description{
A wrapper for the merge function, for interactive use. Prints information and warnings about the merge, then conducts the merge in question (with identical functionality, except that the by variable is set to NULL by default, and no dots parameters can be specified).
}
\details{
The effect of the by variable being set to NULL as a default means that failure to specify a by variable will result in an error. This function will not seek to guess which variables you mean to merge on, i.e. by checking which variables are present in both data frames. The by variable must be specified.
In addition the ability to add ... is gone, because I never use this option (don't know feel like writing the additional code needed to pass this to the merge function).
}
\examples{
x <- data.frame(id = 1:10, xval = 10:1)
y <- data.frame(id = 1:5, yval = 15:11)
SafeMerge(x, y, by = "id")
}
\author{
Olivier Humblet
}
\keyword{merge}

