% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_functions.R
\name{stop_if_many_to_many}
\alias{stop_if_many_to_many}
\title{Generate error if a many-to-many merge is detected.}
\usage{
stop_if_many_to_many(x, y, by.x, by.y)
}
\arguments{
\item{x}{Typically the merged portion of the final merged dataset that came from x. (We only care about the merged portion.)}

\item{y}{Typically the merged portion of the final merged dataset that came from y. (We only care about the merged portion.)}

\item{by.x}{From base::merge.}

\item{by.y}{From base::merge.}
}
\value{
None, throws error.
}
\description{
Generates error if both input data frames have >1 observation for any unique combination of by variable(s).
}
\details{
A many-to-many merge will cause R to generated all combinations of the non-uniquely-matched rows. This is almost never the intended behavior, so it is not allowed by SafeMerge.
}
\examples{
d1 <- data.frame(id = rep(1:2, times = 2:1), v1 = 1:3)
d2 <- data.frame(id = rep(1:2, times = 2:1), v2 = 4:6)
stop_if_many_to_many(d1, d2, "id", "id")
}
\author{
Olivier Humblet
}

