% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_functions.R
\name{stop_if_no_by_vars}
\alias{stop_if_no_by_vars}
\title{Generate error if no by variables were provided.}
\usage{
stop_if_no_by_vars(by, by.x, by.y)
}
\arguments{
\item{by}{The 'by' input to the merge function.}

\item{by.x}{The 'by.x' input to the merge function.}

\item{by.y}{The 'by.y' input to the merge function.}
}
\value{
None, throws error.
}
\description{
Generates error if no by variables were provided.
}
\details{
None.
}
\examples{
stop_if_no_by_vars(NULL, NULL, NULL)
}
\author{
Olivier Humblet
}

