% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_obs_and_ids.R
\name{return_list_ratios_xy}
\alias{return_list_ratios_xy}
\title{Returns a list of the ratios of observations to rows in x and y.}
\usage{
return_list_ratios_xy(x, y, by.x, by.y)
}
\arguments{
\item{x}{Typically the merged portion of the final merged dataset that came from x. (We only care about the merged portion.)}

\item{y}{Typically the merged portion of the final merged dataset that came from y. (We only care about the merged portion.)}

\item{by.x}{by.x from base::merge.}

\item{by.y}{by.y from base::merge.}
}
\value{
Returns a list whose two entries are the ratios of observations to unique by-variable combinations, for the two datasets (i.e. ratio_in_x and ratio_in_y, respectively).
}
\description{
Returns a list of the ratios of observations to rows in both x and y..
}
\details{
None.
}
\examples{
d1 <- data.frame(id = rep(1:2, times = 2:1), v1 = 1:3)
d2 <- data.frame(id = rep(1:2, times = 2:1), v2 = 4:6)
return_list_ratios_xy(d1, d2, "id")
}
\author{
Olivier Humblet
}

