% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_functions.R
\name{stop_if_many_to_many}
\alias{stop_if_many_to_many}
\title{Generate error if a many-to-many merge is detected.}
\usage{
stop_if_many_to_many(x, y, by.x, by.y)
}
\arguments{
\item{x}{Typically the merged portion of the final merged dataset that came from x. (We only care about the merged portion.)}

\item{y}{Typically the merged portion of the final merged dataset that came from y. (We only care about the merged portion.)}

\item{by.x}{The 'by.x' input to the merge function.}

\item{by.y}{The 'by.y' input to the merge function.}
}
\value{
None, throws error.
}
\description{
Generates error if both input data frames have >1 observation for any unique combination of by variable(s).
}
\details{
A many-to-many merge will cause R to generated all combinations of the non-uniquely-matched rows. This is almost never the ntended behavior, so it is not allowed by SafeMerge.
}
\examples{
stop_if_no_by_vars(NULL, NULL, NULL)
}
\author{
Olivier Humblet
}

